+++ packages/snap-account-service/src/SnapAccountService-method-action-types.ts
@@ -49,17 +49,6 @@ export type SnapAccountServiceMigrateAction = {
 };
 
 /**
- * Atomically gets-or-creates the legacy (v1) Snap keyring — the keyring
- * associated with {@link KeyringTypes.snap}.
- *
- * @returns The existing or newly-created Snap keyring instance.
- */
-export type SnapAccountServiceGetLegacySnapKeyringAction = {
-  type: `SnapAccountService:getLegacySnapKeyring`;
-  handler: SnapAccountService['getLegacySnapKeyring'];
-};
-
-/**
  * Handle a message from a Snap.
  *
  * @param snapId - ID of the Snap.
@@ -78,5 +67,4 @@ export type SnapAccountServiceMethodActions =
   | SnapAccountServiceGetSnapsAction
   | SnapAccountServiceEnsureReadyAction
   | SnapAccountServiceMigrateAction
-  | SnapAccountServiceGetLegacySnapKeyringAction
   | SnapAccountServiceHandleKeyringSnapMessageAction;